COMMAND_SHORTCUT_MANAGER - Renamed from SYM_LINK_MANAGER - Refactor to store & load command shortcuts through FileManager instead of user.cfg - COMMAND_SHORTCUT_STORAGE_FILE - Command shortcuts beginning with '/' can be used in both chat and the console
Ability to disable instant chats:
- INSTANT_CHAT_ENABLED
- INSTANT_CHAT_COMMANDS_ONLY

HELPER_HUD:
color configurability:
- HELPER_HUD_COLOR_LABEL
- HELPER_HUD_COLOR_VALUES
- HELPER_HUD_COLOR_ENABLED
- HELPER_HUD_COLOR_DISABLED
Refactored / added additional items

BROWSER_DEFAULT_SORT - added value map
Refactor

+++ src/tron/gCycleMovement.h
@@ -244,16 +244,16 @@ protected:
 
 public:
     REAL lastTurnAttemptTime;
-    REAL lastTurnAttemptDir;
+    int  lastTurnAttemptDir;
     REAL lastTurnTime;
     REAL lastTurnSysTime;
     REAL turnIgnoreTime;
-    REAL lastTurnDir; // 0 = NONE, -1 = left, 1 = right
-    REAL blockTurn; // 0 = NONE, -1 = LEFT, 1 = RIGHT, 2 = BOTH
-    REAL forceTurn; // 0 = NONE, -1 = LEFT, 1 = RIGHT
-    REAL lastBlockedTurn;
+    int  lastTurnDir; // 0 = NONE, -1 = left, 1 = right
+    int  blockTurn;   // 0 = NONE, -1 = LEFT, 1 = RIGHT, 2 = BOTH
+    int  forceTurn;   // 0 = NONE, -1 = LEFT, 1 = RIGHT
+    int  lastBlockedTurn;
     REAL lastBotTurnTime;
-    REAL lastBotTurnDir;
+    int  lastBotTurnDir;
     REAL localCurrentTime;
     // room for accessors
 public:
